Some other warnings, GitHub issue 2382.

+++ control.c
@@ -688,8 +688,8 @@ control_write_pending(struct client *c, struct control_pane *cp, size_t limit)
 		else
 			age = 0;
 		log_debug("%s: %s: output block %zu (age %llu) for %%%u "
-		    "(used %zu/%zu)", __func__, c->name, cb->size, age,
-		    cp->pane, used, limit);
+		    "(used %zu/%zu)", __func__, c->name, cb->size,
+		    (unsigned long long)age, cp->pane, used, limit);
 
 		size = cb->size;
 		if (size > limit - used)
